fix: adjust edit user refactor
