+++ src/mocks/utils/initializeDataDB.ts
@@ -5,24 +5,24 @@ import { intializedTokenData } from "@mocks/themeService/themeService.mock";
 import { DocumentsServiceMock } from "@mocks/privileges/documents/DocumentsServiceMock.mock";
 import { clientServerMock } from "@mocks/privileges/client-server/client-serverServiceMock.mock";
 import { buttonOptionsMock } from "@mocks/privileges/button/buttonOptionsMock.mock";
-import { webOptionsMock } from "@src/mocks/privileges/web/webOptionsMock.mock";
+import { webOptionsMock } from "@mocks/privileges/web/webOptionsMock.mock";
 import { linixUseCases } from "@mocks/privileges/linixUseCases/LinixUseCases.mock";
 import { mockRoles } from "@mocks/privileges/roles/Roles.mock";
 import { mockActions } from "@mocks/privileges/actions/Actions.mock";
 import { MockPositions } from "@mocks/privileges/positions/Positions.mock";
-import { userEntriesDataMock } from "../apps/privileges/users/users.mock";
-import { invitationEntriesDataMock } from "../apps/privileges/invitations/invitations.mock";
-import { branchesFormEditUser } from "../apps/privileges/users/branchesForm.mock";
-import { projectsFormEditUser } from "../apps/privileges/users/projectsForm.mock";
-import { eventsFormEditUser } from "../apps/privileges/users/eventsForm.mock";
-import { aidBudgetsFormEditUser } from "../apps/privileges/users/aidBudgetsForm.mock";
-import { payrollsFormEditUser } from "../apps/privileges/users/payrollsForm.mock";
-import { branchesFormInvitation } from "../apps/privileges/invitations/branchesForm.mock";
-import { projectsFormInvitation } from "../apps/privileges/invitations/projectsForm.mock";
-import { eventsFormInvitation } from "../apps/privileges/invitations/eventsForm.mock";
-import { aidBudgetsFormInvitation } from "../apps/privileges/invitations/aidBudgetsForm.mock";
-import { payrollsFormInvitation } from "../apps/privileges/invitations/payrollsForm.mock";
-import { invitationUserEntriesDataMock } from "../apps/privileges/invitations/invitationUsers.mock";
+import { userEntriesDataMock } from "@mocks/apps/privileges/users/users.mock";
+import { invitationEntriesDataMock } from "@mocks/apps/privileges/invitations/invitations.mock";
+import { branchesFormEditUser } from "@mocks/apps/privileges/users/branchesForm.mock";
+import { projectsFormEditUser } from "@mocks/apps/privileges/users/projectsForm.mock";
+import { eventsFormEditUser } from "@mocks/apps/privileges/users/eventsForm.mock";
+import { aidBudgetsFormEditUser } from "@mocks/apps/privileges/users/aidBudgetsForm.mock";
+import { payrollsFormEditUser } from "@mocks/apps/privileges/users/payrollsForm.mock";
+import { branchesFormInvitation } from "@mocks/apps/privileges/invitations/branchesForm.mock";
+import { projectsFormInvitation } from "@mocks/apps/privileges/invitations/projectsForm.mock";
+import { eventsFormInvitation } from "@mocks/apps/privileges/invitations/eventsForm.mock";
+import { aidBudgetsFormInvitation } from "@mocks/apps/privileges/invitations/aidBudgetsForm.mock";
+import { payrollsFormInvitation } from "@mocks/apps/privileges/invitations/payrollsForm.mock";
+import { invitationUserEntriesDataMock } from "@mocks/apps/privileges/invitations/invitationUsers.mock";
 
 export function initializeDataDB() {
   localforage.clear();
