+++ src/pages/privileges/outlets/users/tabs/users/config/dataUsers.config.tsx
@@ -1,14 +1,14 @@
+import { Link } from "react-router-dom";
+import { MdModeEdit } from "react-icons/md";
+
+import { Icon } from "@inube/design-system";
+import { IGeneralInformationUsersForm } from "@src/services/users/users.types";
 import { deleteItemData } from "@mocks/utils/dataMock.service";
 
 import { DeleteLinixUsers } from "../DeleteModal";
 import { deleteLinixUsersModal } from "../DeleteModal/config/deleteLinixUsers.config";
 import { ActivateUsers } from "../ActivateFormOptions";
 import { activateUsersModal } from "../ActivateFormOptions/config/activateUsers.config";
-import { Icon } from "@inube/design-system";
-
-import { Link } from "react-router-dom";
-import { MdModeEdit } from "react-icons/md";
-import { IGeneralInformationUsersForm } from "@src/services/users/users.types";
 
 export const actionsConfigUsers = (
   smallScreen: boolean,
