+++ src/services/users/users.types.ts
@@ -1,4 +1,4 @@
-import { IAssignmentFormEntry } from "@src/pages/privileges/outlets/users/types/forms.types";
+import { IAssignmentFormEntry } from "@pages/privileges/outlets/users/types/forms.types";
 
 export interface IGeneralInformationUsersForm {
   k_Usuari: string;
